fix(TF-72): fetch fail 오류 수정
